requirements.txt
requests
xmpppy
